+++ engine/src/lib.rs
@@ -4,7 +4,13 @@
     all(feature = "log", target_arch = "wasm32"),
     feature(panic_info_message)
 )]
-#![deny(clippy::as_conversions)]
+#![deny(clippy::pedantic, clippy::nursery)]
+#![allow(
+    clippy::missing_errors_doc,
+    clippy::missing_panics_doc,
+    clippy::module_name_repetitions,
+    clippy::unreadable_literal
+)]
 
 use aurora_engine_types::parameters::PromiseCreateArgs;
 
@@ -71,6 +77,7 @@ pub unsafe fn on_alloc_error(_: core::alloc::Layout) -> ! {
 mod contract {
     use ::function_name::named;
     use borsh::{BorshDeserialize, BorshSerialize};
+    use parameters::SetOwnerArgs;
 
     use crate::connector::{self, EthConnectorContract};
     use crate::engine::{self, Engine};
@@ -81,7 +88,7 @@ mod contract {
         self, CallArgs, DeployErc20TokenArgs, GetErc20FromNep141CallArgs, GetStorageAtArgs,
         InitCallArgs, IsUsedProofCallArgs, NEP141FtOnTransferArgs, NewCallArgs,
         PauseEthConnectorCallArgs, PausePrecompilesCallArgs, ResolveTransferCallArgs,
-        SetContractDataCallArgs, StorageDepositCallArgs, StorageWithdrawCallArgs,
+        SetContractDataCallArgs, StorageDepositCallArgs, StorageWithdrawCallArgs, SubmitArgs,
         TransferCallCallArgs, ViewCallArgs,
     };
     #[cfg(feature = "evm_bully")]
@@ -134,7 +141,7 @@ mod contract {
             [0; 32],
         );
 
-        state::set_state(&mut io, state).sdk_unwrap();
+        state::set_state(&mut io, &state).sdk_unwrap();
         hashchain::set_state(&mut io, blockchain_hashchain).sdk_unwrap();
     }
 
@@ -142,11 +149,9 @@ mod contract {
     #[no_mangle]
     pub extern "C" fn get_version() {
         let mut io = Runtime;
-        let version = match option_env!("NEAR_EVM_VERSION") {
-            Some(v) => v.as_bytes(),
-            None => include_bytes!("../../VERSION"),
-        };
-        io.return_output(version)
+        let version = option_env!("NEAR_EVM_VERSION")
+            .map_or(&include_bytes!("../../VERSION")[..], str::as_bytes);
+        io.return_output(version);
     }
 
     /// Get owner account id for this contract.
@@ -155,6 +160,21 @@ mod contract {
         let mut io = Runtime;
         let state = state::get_state(&io).sdk_unwrap();
         io.return_output(state.owner_id.as_bytes());
+    }
+
+    /// Set owner account id for this contract.
+    #[no_mangle]
+    pub extern "C" fn set_owner() {
+        let mut io = Runtime;
+        let mut state = state::get_state(&io).sdk_unwrap();
+        require_owner_only(&state, &io.predecessor_account_id());
+        let args: SetOwnerArgs = io.read_input_borsh().sdk_unwrap();
+        if state.owner_id == args.new_owner {
+            sdk::panic_utf8(errors::ERR_SAME_OWNER);
+        } else {
+            state.owner_id = args.new_owner;
+            state::set_state(&mut io, &state).sdk_unwrap();
+        }
     }
 
     /// Get bridge prover id for this contract.
@@ -169,15 +189,14 @@ mod contract {
     #[no_mangle]
     pub extern "C" fn get_chain_id() {
         let mut io = Runtime;
-        io.return_output(&state::get_state(&io).sdk_unwrap().chain_id)
+        io.return_output(&state::get_state(&io).sdk_unwrap().chain_id);
     }
 
     #[no_mangle]
     pub extern "C" fn get_upgrade_index() {
         let mut io = Runtime;
-        let state = state::get_state(&io).sdk_unwrap();
         let index = internal_get_upgrade_index();
-        io.return_output(&(index + state.upgrade_delay_blocks).to_le_bytes())
+        io.return_output(&index.to_le_bytes());
     }
 
     /// Stage new code for deployment.
@@ -185,12 +204,12 @@ mod contract {
     pub extern "C" fn stage_upgrade() {
         let mut io = Runtime;
         let state = state::get_state(&io).sdk_unwrap();
-        let block_height = io.block_height();
+        let delay_block_height = io.block_height() + state.upgrade_delay_blocks;
         require_owner_only(&state, &io.predecessor_account_id());
         io.read_input_and_store(&bytes_to_key(KeyPrefix::Config, CODE_KEY));
         io.write_storage(
             &bytes_to_key(KeyPrefix::Config, CODE_STAGE_KEY),
-            &block_height.to_le_bytes(),
+            &delay_block_height.to_le_bytes(),
         );
     }
 
@@ -201,7 +220,7 @@ mod contract {
         let state = state::get_state(&io).sdk_unwrap();
         require_owner_only(&state, &io.predecessor_account_id());
         let index = internal_get_upgrade_index();
-        if io.block_height() <= index + state.upgrade_delay_blocks {
+        if io.block_height() <= index {
             sdk::panic_utf8(errors::ERR_NOT_ALLOWED_TOO_EARLY);
         }
         Runtime::self_deploy(&bytes_to_key(KeyPrefix::Config, CODE_KEY));
@@ -211,6 +230,7 @@ mod contract {
     /// to make any necessary changes to the state such that it aligns with the newly deployed
     /// code.
     #[no_mangle]
+    #[allow(clippy::missing_const_for_fn)]
     pub extern "C" fn state_migration() {
         // TODO: currently we don't have migrations
     }
@@ -244,7 +264,7 @@ mod contract {
         let authorizer: pausables::EngineAuthorizer = engine::get_authorizer();
 
         if !authorizer.is_authorized(&io.predecessor_account_id()) {
-            sdk::panic_utf8("ERR_UNAUTHORIZED".as_bytes());
+            sdk::panic_utf8(b"ERR_UNAUTHORIZED");
         }
 
         let input = io.read_input();
@@ -305,8 +325,21 @@ mod contract {
         let input = io.read_input().to_vec();
         let args = CallArgs::deserialize(&input).sdk_expect(errors::ERR_BORSH_DESERIALIZE);
         let current_account_id = io.current_account_id();
+        let predecessor_account_id = io.predecessor_account_id();
+
+        // During the XCC flow the Engine will call itself to move wNEAR
+        // to the user's sub-account. We do not want this move to happen
+        // if prior promises in the flow have failed.
+        if current_account_id == predecessor_account_id {
+            let check_promise: Result<(), &[u8]> = match io.promise_result_check() {
+                Some(true) | None => Ok(()),
+                Some(false) => Err(b"ERR_CALLBACK_OF_FAILED_PROMISE"),
+            };
+            check_promise.sdk_unwrap();
+        }
+
         let mut engine = Engine::new(
-            predecessor_address(&io.predecessor_account_id()),
+            predecessor_address(&predecessor_account_id),
             current_account_id,
             io,
             &io,
@@ -325,7 +358,7 @@ mod contract {
     }
 
     /// Process signed Ethereum transaction.
-    /// Must match CHAIN_ID to make sure it's signed for given chain vs replayed from another chain.
+    /// Must match `CHAIN_ID` to make sure it's signed for given chain vs replayed from another chain.
     #[no_mangle]
     #[named]
     pub extern "C" fn submit() {
@@ -334,10 +367,14 @@ mod contract {
         let current_account_id = io.current_account_id();
         let state = state::get_state(&io).sdk_unwrap();
         let relayer_address = predecessor_address(&io.predecessor_account_id());
+        let args = SubmitArgs {
+            tx_data: input.clone(),
+            ..Default::default()
+        };
         let result = engine::submit(
             io,
             &io,
-            &input,
+            &args,
             state,
             current_account_id,
             relayer_address,
@@ -348,6 +385,36 @@ mod contract {
 
         if let Ok(output) = &result {
             update_hashchain(&mut io, function_name!(), &input, output);
+        }
+
+        result.sdk_process();
+    }
+
+    /// Analog of the `submit` function, but waits for the `SubmitArgs` structure rather than
+    /// the array of bytes representing the transaction.
+    #[no_mangle]
+    #[named]
+    pub extern "C" fn submit_with_args() {
+        let mut io = Runtime;
+        let input = io.read_input();
+        let args: SubmitArgs = input.to_value().sdk_unwrap();
+        let current_account_id = io.current_account_id();
+        let state = state::get_state(&io).sdk_unwrap();
+        let relayer_address = predecessor_address(&io.predecessor_account_id());
+        let result = engine::submit(
+            io,
+            &io,
+            &args,
+            state,
+            current_account_id,
+            relayer_address,
+            &mut Runtime,
+        );
+
+        let result = result.map(|res| res.try_to_vec().sdk_expect(errors::ERR_SERIALIZE));
+
+        if let Ok(output) = &result {
+            update_hashchain(&mut io, function_name!(), &input.to_vec(), output);
         }
 
         result.sdk_process();
@@ -399,10 +466,11 @@ mod contract {
     #[named]
     pub extern "C" fn factory_update_address_version() {
         let mut io = Runtime;
+        // The function is only set to be private, otherwise callback error will happen.
         io.assert_private_call().sdk_unwrap();
-        let check_deploy: Result<(), &[u8]> = match io.promise_result(0) {
-            Some(PromiseResult::Successful(_)) => Ok(()),
-            Some(_) => Err(b"ERR_ROUTER_DEPLOY_FAILED"),
+        let check_deploy: Result<(), &[u8]> = match io.promise_result_check() {
+            Some(true) => Ok(()),
+            Some(false) => Err(b"ERR_ROUTER_DEPLOY_FAILED"),
             None => Err(b"ERR_ROUTER_UPDATE_NOT_CALLBACK"),
         };
         check_deploy.sdk_unwrap();
@@ -413,7 +481,7 @@ mod contract {
         update_hashchain(&mut io, function_name!(), &input.to_vec(), &[]);
     }
 
-    /// Sets the address for the wNEAR ERC-20 contract. This contract will be used by the
+    /// Sets the address for the `wNEAR` ERC-20 contract. This contract will be used by the
     /// cross-contract calls feature to have users pay for their NEAR transactions.
     #[no_mangle]
     #[named]
@@ -518,7 +586,7 @@ mod contract {
             let args: RefundCallArgs = input.to_value().sdk_unwrap();
             let state = state::get_state(&io).sdk_unwrap();
             let refund_result =
-                engine::refund_on_error(io, &io, state, args, &mut Runtime).sdk_unwrap();
+                engine::refund_on_error(io, &io, state, &args, &mut Runtime).sdk_unwrap();
 
             if !refund_result.status.is_ok() {
                 sdk::panic_utf8(errors::ERR_REFUND_FAILURE);
@@ -552,7 +620,7 @@ mod contract {
             .sdk_unwrap();
         let block_hash =
             crate::engine::compute_block_hash(chain_id, block_height, account_id.as_bytes());
-        io.return_output(block_hash.as_bytes())
+        io.return_output(block_hash.as_bytes());
     }
 
     #[no_mangle]
@@ -589,7 +657,7 @@ mod contract {
         let mut io = Runtime;
         let address = io.read_input_arr20().sdk_unwrap();
         let code = engine::get_code(&io, &Address::from_array(address));
-        io.return_output(&code)
+        io.return_output(&code);
     }
 
     #[no_mangle]
@@ -597,7 +665,7 @@ mod contract {
         let mut io = Runtime;
         let address = io.read_input_arr20().sdk_unwrap();
         let balance = engine::get_balance(&io, &Address::from_array(address));
-        io.return_output(&balance.to_bytes())
+        io.return_output(&balance.to_bytes());
     }
 
     #[no_mangle]
@@ -605,7 +673,7 @@ mod contract {
         let mut io = Runtime;
         let address = io.read_input_arr20().sdk_unwrap();
         let nonce = engine::get_nonce(&io, &Address::from_array(address));
-        io.return_output(&u256_to_arr(&nonce))
+        io.return_output(&u256_to_arr(&nonce));
     }
 
     #[no_mangle]
@@ -615,7 +683,7 @@ mod contract {
         let address = args.address;
         let generation = engine::get_generation(&io, &address);
         let value = engine::get_storage(&io, &args.address, &H256(args.key), generation);
-        io.return_output(&value.0)
+        io.return_output(&value.0);
     }
 
     ///
@@ -630,17 +698,17 @@ mod contract {
         require_owner_only(&state, &io.predecessor_account_id());
         let args: BeginChainArgs = io.read_input_borsh().sdk_unwrap();
         state.chain_id = args.chain_id;
-        state::set_state(&mut io, state).sdk_unwrap();
+        state::set_state(&mut io, &state).sdk_unwrap();
         // set genesis block balances
         for account_balance in args.genesis_alloc {
             engine::set_balance(
                 &mut io,
                 &account_balance.address,
                 &crate::prelude::Wei::new(U256::from(account_balance.balance)),
-            )
+            );
         }
         // return new chain ID
-        io.return_output(&state::get_state(&io).sdk_unwrap().chain_id)
+        io.return_output(&state::get_state(&io).sdk_unwrap().chain_id);
     }
 
     #[cfg(feature = "evm_bully")]
@@ -658,13 +726,17 @@ mod contract {
     pub extern "C" fn new_eth_connector() {
         let mut io = Runtime;
         // Only the owner can initialize the EthConnector
-        io.assert_private_call().sdk_unwrap();
+        let is_private = io.assert_private_call();
+        if is_private.is_err() {
+            let state = state::get_state(&io).sdk_unwrap();
+            require_owner_only(&state, &io.predecessor_account_id());
+        }
 
         let input = io.read_input();
         let args: InitCallArgs = input.to_value().sdk_unwrap();
         let owner_id = io.current_account_id();
 
-        EthConnectorContract::create_contract(io, owner_id, args).sdk_unwrap();
+        EthConnectorContract::create_contract(io, &owner_id, args).sdk_unwrap();
 
         update_hashchain(&mut io, function_name!(), &input.to_vec(), &[]);
     }
@@ -674,7 +746,11 @@ mod contract {
     pub extern "C" fn set_eth_connector_contract_data() {
         let mut io = Runtime;
         // Only the owner can set the EthConnector contract data
-        io.assert_private_call().sdk_unwrap();
+        let is_private = io.assert_private_call();
+        if is_private.is_err() {
+            let state = state::get_state(&io).sdk_unwrap();
+            require_owner_only(&state, &io.predecessor_account_id());
+        }
 
         let input = io.read_input();
         let args: SetContractDataCallArgs = input.to_value().sdk_unwrap();
@@ -694,7 +770,7 @@ mod contract {
         let predecessor_account_id = io.predecessor_account_id();
         let result = EthConnectorContract::init_instance(io)
             .sdk_unwrap()
-            .withdraw_eth_from_near(&current_account_id, &predecessor_account_id, args)
+            .withdraw_eth_from_near(&current_account_id, &predecessor_account_id, &args)
             .sdk_unwrap();
         let result_bytes = result.try_to_vec().sdk_expect(errors::ERR_SERIALIZE);
         // We intentionally do not go through the `io` struct here because we must bypass
@@ -786,7 +862,7 @@ mod contract {
 
         let is_used_proof = EthConnectorContract::init_instance(io)
             .sdk_unwrap()
-            .is_used_proof(args.proof);
+            .is_used_proof(&args.proof);
         let res = is_used_proof.try_to_vec().unwrap();
         io.return_output(&res[..]);
     }
@@ -823,7 +899,7 @@ mod contract {
             .sdk_unwrap();
         EthConnectorContract::init_instance(io)
             .sdk_unwrap()
-            .ft_balance_of(args);
+            .ft_balance_of(&args);
     }
 
     #[no_mangle]
@@ -832,7 +908,7 @@ mod contract {
         let args: parameters::BalanceOfEthCallArgs = io.read_input().to_value().sdk_unwrap();
         EthConnectorContract::init_instance(io)
             .sdk_unwrap()
-            .ft_balance_of_eth_on_aurora(args)
+            .ft_balance_of_eth_on_aurora(&args)
             .sdk_unwrap();
     }
 
@@ -848,7 +924,7 @@ mod contract {
             .sdk_unwrap();
         EthConnectorContract::init_instance(io)
             .sdk_unwrap()
-            .ft_transfer(&predecessor_account_id, args)
+            .ft_transfer(&predecessor_account_id, &args)
             .sdk_unwrap();
 
         update_hashchain(&mut io, function_name!(), &input, &[]);
@@ -870,7 +946,7 @@ mod contract {
 
         EthConnectorContract::init_instance(io)
             .sdk_unwrap()
-            .ft_resolve_transfer(args, promise_result);
+            .ft_resolve_transfer(&args, promise_result);
 
         update_hashchain(&mut io, function_name!(), &input.to_vec(), &[]);
     }
@@ -961,7 +1037,7 @@ mod contract {
         let predecessor_account_id = io.predecessor_account_id();
         EthConnectorContract::init_instance(io)
             .sdk_unwrap()
-            .storage_withdraw(&predecessor_account_id, args)
+            .storage_withdraw(&predecessor_account_id, &args)
             .sdk_unwrap();
 
         update_hashchain(&mut io, function_name!(), &input, &[]);
@@ -976,7 +1052,7 @@ mod contract {
                 .sdk_unwrap();
         EthConnectorContract::init_instance(io)
             .sdk_unwrap()
-            .storage_balance_of(args)
+            .storage_balance_of(&args);
     }
 
     #[no_mangle]
@@ -993,13 +1069,16 @@ mod contract {
     #[named]
     pub extern "C" fn set_paused_flags() {
         let mut io = Runtime;
-        io.assert_private_call().sdk_unwrap();
-
+        let is_private = io.assert_private_call();
+        if is_private.is_err() {
+            let state = state::get_state(&io).sdk_unwrap();
+            require_owner_only(&state, &io.predecessor_account_id());
+        }
         let input = io.read_input();
         let args: PauseEthConnectorCallArgs = input.to_value().sdk_unwrap();
         EthConnectorContract::init_instance(io)
             .sdk_unwrap()
-            .set_paused_flags(args);
+            .set_paused_flags(&args);
 
         update_hashchain(&mut io, function_name!(), &input.to_vec(), &[]);
     }
@@ -1042,7 +1121,7 @@ mod contract {
         let mut io = Runtime;
         let metadata: FungibleTokenMetadata = connector::get_metadata(&io).unwrap_or_default();
         let bytes = serde_json::to_vec(&metadata).unwrap_or_default();
-        io.return_output(&bytes)
+        io.return_output(&bytes);
     }
 
     #[cfg(feature = "integration-test")]
@@ -1059,7 +1138,7 @@ mod contract {
     #[no_mangle]
     pub extern "C" fn mint_account() {
         use crate::connector::ZERO_ATTACHED_BALANCE;
-        use crate::prelude::{NEP141Wei, ToString, U256};
+        use crate::prelude::{NEP141Wei, U256};
         use evm::backend::ApplyBackend;
         const GAS_FOR_VERIFY: NearGas = NearGas::new(20_000_000_000_000);
         const GAS_FOR_FINISH: NearGas = NearGas::new(50_000_000_000_000);
@@ -1096,14 +1175,14 @@ mod contract {
         };
         let verify_call = aurora_engine_types::parameters::PromiseCreateArgs {
             target_account_id: aurora_account_id.clone(),
-            method: "verify_log_entry".to_string(),
+            method: crate::prelude::String::from("verify_log_entry"),
             args: crate::prelude::Vec::new(),
             attached_balance: ZERO_ATTACHED_BALANCE,
             attached_gas: GAS_FOR_VERIFY,
         };
         let finish_call = aurora_engine_types::parameters::PromiseCreateArgs {
             target_account_id: aurora_account_id,
-            method: "finish_deposit".to_string(),
+            method: crate::prelude::String::from("finish_deposit"),
             args: args.try_to_vec().unwrap(),
             attached_balance: ZERO_ATTACHED_BALANCE,
             attached_gas: GAS_FOR_FINISH,