+++ CHANGES.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- New variant of submit function `submit_with_args` which accepts additional arguments along with the transaction.
+
 ## [2.8.0] 2022-11-15
 
 ### Added